moved graph generation to backend express
