+++ client/actions/getGraph.js
@@ -1,0 +1,5 @@
+import axios from 'axios';
+
+export function getGraphData(params) {
+	return axios.get('/api/getGraphData', {params});
+}
