+++ package.json
@@ -21,6 +21,7 @@
     "@babel/core": "^7.0.0-beta.42",
     "@babel/preset-env": "^7.0.0-beta.42",
     "@babel/preset-react": "^7.0.0-beta.42",
+    "@material-ui/core": "^4.11.0",
     "@nivo/pie": "^0.57.0",
     "autoprefixer": "^8.2.0",
     "axios": "^0.18.1",
